fix: create class on separated files
