+++ SQLGame/Assets/Scripts/Clues/Clue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e94fc4c3dc0485741a2ecee20d9d3ec6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
