+++ SQLGame/Assets/Scripts/Clues/ClueSolution.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 405dba3d0611a6d4da3492f06d32227e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
