+++ SQLGame/Assets/Scripts/Clues/SolutionPart.cs
@@ -1,0 +1,13 @@
+public class SolutionPart
+{
+    public string description;
+
+    public virtual bool IsFound() 
+    { 
+        return false; 
+    } //to be implemented in subclasses
+
+    public virtual void UpdateController() { }
+
+    public ClueController clueController;
+}
