+++ SQLGame/Assets/Scripts/Clues/SolutionPart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0e8347a2391e5f43af9aa75764f2d3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
